ref: logic to detect if the poster is default moved to the datasource handler

+++ src/services/datasource.js
@@ -15,6 +15,7 @@ const datasource = {
   },
 
   async removeFile(filePath) {
+    if (poster === "/poster/default.png") return; //early return on default poster
     const file = path.join(__dirname, "../public", filePath);
     try {
       await fs.unlink(file);
